+++ app/Providers/AppServiceProvider.php
@@ -9,6 +9,7 @@ use App\Models\Taxonomy;
 use App\Policies\MenuPolicy;
 use App\Policies\PostPolicy;
 use App\Policies\TaxonomyPolicy;
+use Illuminate\Auth\Notifications\ResetPassword;
 use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\View;
 use Illuminate\Support\ServiceProvider;
@@ -25,6 +26,10 @@ class AppServiceProvider extends ServiceProvider
         Gate::policy(Menu::class, MenuPolicy::class);
         Gate::policy(Post::class, PostPolicy::class);
         Gate::policy(Taxonomy::class, TaxonomyPolicy::class);
+
+        ResetPassword::createUrlUsing(function ($user, string $token) {
+            return route('admin.password.reset', ['token' => $token, 'email' => $user->getEmailForPasswordReset()]);
+        });
 
         View::composer('layouts.public', function ($view) {
             $siteLogo = Setting::get('site_logo');